Fix AppErrorDictionary JC-551
